4545 client - icons in Skills left sidebar + re-register parent post-turn callback on view change
Sidebar icons:
- Render a small colored square + Zap icon next to each skill name in
  AiSkillsLeftSidebar, matching the main list. The color is derived from
  a hash of the skill id so the same skill always renders the same
  color across surfaces.
- Extract getSkillColor (previously inline in useAiSkillListItem) to a
  shared util so both list and sidebar consume the same mapping.
- Sidebar icon is size-5/size-3 (smaller than the list-row size-8/size-4)
  to fit the denser sidebar row.

Parent post-turn registration:
- Add skillsView to AiSkills' Source.SKILLS register effect deps.
  AiSkillCreateWithAi overwrites the parent's invalidate-only callback
  while mounted; without the dep change, when the page unmounted the
  parent's callback was already removed (cleanup happened via
  AiSkillCreateWithAi's cleanup) and never reinstated. Triggering
  re-registration on skillsView change reinstalls it as the view
  transitions back out of createWithAi.
