+++ client/src/pages/automation/ai/skills/hooks/useAiSkillListItem.ts
@@ -1,19 +1,8 @@
 import {useAiSkillsStore} from '@/pages/automation/ai/skills/stores/useAiSkillsStore';
+import getSkillColor from '@/pages/automation/ai/skills/utils/getSkillColor';
 import {AiSkill} from '@/shared/middleware/graphql';
 import {useCallback, useState} from 'react';
 import {useNavigate} from 'react-router-dom';
-
-const SKILL_COLORS = ['bg-blue-500', 'bg-green-500', 'bg-purple-500', 'bg-orange-500', 'bg-pink-500', 'bg-teal-500'];
-
-function getSkillColor(skillId: string): string {
-    let hash = 0;
-
-    for (let index = 0; index < skillId.length; index++) {
-        hash = skillId.charCodeAt(index) + ((hash << 5) - hash);
-    }
-
-    return SKILL_COLORS[Math.abs(hash) % SKILL_COLORS.length];
-}
 
 interface UseAiSkillListItemPropsI {
     deleteSkill: (id: string) => Promise<void>;
